top_camera_final/requirements.txt
ultralytics 
pyzbar 
qreader 
kivy 
kivymd 
requests 
python-socketio 
websocket-client